apps/users/serializers.py
# File: users/serializers.py
from rest_framework import serializers
from .models import User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['user_id', 'email', 'phone_number', 'date_joined', 'is_active', 'is_staff']
