requirements.txt
asgiref==3.7.2
awscli==1.29.57
bcrypt==4.0.1
botocore==1.31.57
cffi==1.16.0
colorama==0.4.4
cryptography==41.0.5
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
docutils==0.16
jmespath==1.0.1
loguru==0.7.2
markdown-it-py==3.0.0
mdurl==0.1.2
paramiko==3.3.1
psycopg2==2.9.9
pyasn1==0.5.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.5.0
PyNaCl==1.5.0
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyup==0.1.4
PyYAML==6.0.1
rich==13.6.0
rsa==4.7.2
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.16
